Added sphere with uv

+++ include/easy_pbr/Frame.h
@@ -64,7 +64,7 @@ public:
     bool is_keyframe=false; //if it is keyframe we would need to create seeds
 
     std::shared_ptr<Mesh> create_frustum_mesh(float scale_multiplier=1.0, bool show_texture=true, const int texture_max_size=256) const;
-    Frame subsample(const int subsample_factor);
+    Frame subsample(const float subsample_factor);
     // void rotate_y_axis(const float rads );
     // Mesh backproject_depth() const;
     // Mesh assign_color(Mesh& cloud);
